Fix Approver dependency in acceptance tests + fix and improve flight service
Also fix missing .war at image startup

+++ services/flight/src/main/java/flight/data/filter/MaxTravelTimeFB.java
@@ -13,6 +13,6 @@ public class MaxTravelTimeFB implements FilterBuilder<Bson> {
             return null;
         }
 
-        return lte("duration", bson.getLong("maxTravelTime"));
+        return lte("duration", bson.getInteger("maxTravelTime").longValue());
     }
 }
